# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  handler.LogoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
  handler.MeResponse:
    properties:
      id:
        type: string
      login:
        type: string
      role:
        type: string
    type: object
  handler.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  handler.RegisterRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  handler.TokensResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: JWT-based authentication service using RS256 with refresh token rotation.
  title: auth-mytierlist API
  version: "1.0"
paths:
  /.well-known/jwks.json:
    get:
      description: Returns the RSA public key set used to verify access tokens (RS256)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              keys:
                items:
                  properties:
                    alg:
                      type: string
                    e:
                      type: string
                    kid:
                      type: string
                    kty:
                      type: string
                    "n":
                      type: string
                    use:
                      type: string
                  type: object
                type: array
            type: object
      summary: JSON Web Key Set
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with login and password, receive access + refresh
        tokens
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke the provided refresh token; idempotent — safe to call even
        if already logged out
      parameters:
      - description: Refresh token to revoke
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: Returns profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MeResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access + refresh token
        pair (rotation)
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Token invalid, expired, or already used
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account; returns access + refresh tokens on success
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.TokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Login already taken
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer eyJ..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"